HZN-1307: Move features/persistence modules to features/distributed
