+++ features/distributed/config/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,18 @@
+<blueprint
+        xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="
+        http://www.osgi.org/xmlns/blueprint/v1.0.0
+        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+">
+    <reference id="restClient" interface="org.opennms.distributed.core.api.RestClient" availability="mandatory"/>
+
+    <bean id="configProxyAnnotationBeanPostProcessor" class="org.opennms.features.distributed.config.ConfigProxyAnnotationBeanPostProcessor">
+        <property name="restClient" ref="restClient" />
+    </bean>
+    <service ref="configProxyAnnotationBeanPostProcessor" interface="org.springframework.beans.factory.config.BeanPostProcessor">
+        <service-properties>
+            <entry key="name" value="configProxyAnnotationBeanPostProcessor"/>
+        </service-properties>
+    </service>
+</blueprint>
