+++ features/distributed/dao-impl/src/main/resources/META-INF/services/javax.validation.spi.ValidationProvider
