+++ features/distributed/pom.xml
@@ -17,6 +17,10 @@
   <description>All features shared by distributed containers such as minion or sentinel</description>
   <packaging>pom</packaging>
   <modules>
+    <module>config</module>
     <module>core</module>
+    <module>dao-api</module>
+    <module>dao-impl</module>
+    <module>datasource</module>
   </modules>
 </project>
