--- features/persistence/dao-api/pom.xml
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>org.opennms.features</groupId>
-        <artifactId>persistence-parent</artifactId>
-        <version>22.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>org.opennms.features.persistence</groupId>
-    <artifactId>org.opennms.features.persistence.dao-api</artifactId>
-    <packaging>bundle</packaging>
-    <name>OpenNMS :: Features :: Persistence :: DAO :: API</name>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <extensions>true</extensions>
-                <configuration>
-                    <instructions>
-                        <Bundle-RequiredExecutionEnvironment>JavaSE-1.8</Bundle-RequiredExecutionEnvironment>
-                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <!-- TODO MVR when merging with features/drift add "org.opennms.netmgt.flows.classification.persistence.api" -->
-                        <Export-Package>
-                            org.opennms.netmgt.dao.util,
-                            org.opennms.netmgt.dao.api,
-                            org.opennms.netmgt.bsm.persistence.api,
-                            org.opennms.netmgt.bsm.persistence.api.functions,
-                            org.opennms.netmgt.bsm.persistence.api.functions.reduce,
-                            org.opennms.netmgt.bsm.persistence.api.functions.map
-                        </Export-Package>
-                        <!-- TODO MVR when merging with features/drift add "org.opennms.netmgt.flows.classification.persistence.api" -->
-                        <Import-Package>
-                            com.google.common.base,
-                            com.google.common.collect,
-
-                            javax.persistence,
-                            javax.validation.constraints,
-
-                            org.slf4j,
-
-                            org.apache.commons.lang.builder,
-
-                            javassist.util.proxy,
-
-                            org.hibernate,
-                            org.hibernate.criterion,
-                            org.hibernate.proxy,
-
-                            org.springframework.dao,
-                            org.springframework.stereotype,
-
-                            org.opennms.core.criteria,
-                            org.opennms.core.utils,
-                            org.opennms.core.xml,
-                            org.opennms.features.reporting.model,
-                            org.opennms.netmgt.bsm.persistence.api.functions.map,
-                            org.opennms.netmgt.bsm.persistence.api.functions.reduce,
-                            org.opennms.netmgt.config.ackd,
-                            org.opennms.netmgt.config.api,
-                            org.opennms.netmgt.config.javamail,
-                            org.opennms.netmgt.config.microblog,
-                            org.opennms.netmgt.config.provisiond,
-                            org.opennms.netmgt.config.reportd,
-                            org.opennms.netmgt.config.reporting,
-                            org.opennms.netmgt.config.siteStatusViews,
-                            org.opennms.netmgt.config.statsd.model,
-                            org.opennms.netmgt.config.surveillanceViews,
-                            org.opennms.netmgt.config.tl1d,
-                            org.opennms.netmgt.dao.api,
-                            org.opennms.netmgt.dao.util,
-                            org.opennms.netmgt.events.api,
-                            org.opennms.netmgt.model,
-                            org.opennms.netmgt.model.alarm,
-                            org.opennms.netmgt.model.minion,
-                            org.opennms.netmgt.model.monitoringLocations,
-                            org.opennms.netmgt.model.outage,
-                            org.opennms.netmgt.model.topology,
-                            org.opennms.netmgt.xml.event
-                        </Import-Package>
-                        <!-- TODO MVR when merging with features/drift add "org.opennms.netmgt.flows.classification.persistence.api" -->
-                        <Private-Package>
-                            org.opennms.netmgt.dao.util,
-                            org.opennms.netmgt.dao.api,
-                            org.opennms.netmgt.bsm.persistence.api,
-                            org.opennms.netmgt.bsm.persistence.api.functions.reduce,
-                            org.opennms.netmgt.bsm.persistence.api.functions.map
-                        </Private-Package>
-                    </instructions>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>2.4.3</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                        <configuration>
-                            <artifactSet>
-                                <includes>
-                                    <include>org.opennms:opennms-dao-api</include>
-                                    <include>org.opennms.features.bsm:org.opennms.features.bsm.persistence.api</include>
-                                    <!-- TODO MVR comment in, when merging with features/drift -->
-                                    <!--<include>org.opennms.features.flows.classification.persistence:org.opennms.features.flows.classification.persistence.api</include>-->
-                                </includes>
-                            </artifactSet>
-                            <transformers>
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.XmlAppendingTransformer">
-                                    <resource>META-INF/opennms/applicationContext-dao.xml</resource>
-                                    <resource>META-INF/opennms/applicationContext-osgi.xml</resource>
-                                    <resource>META-INF/opennms/component-dao.xml</resource>
-                                </transformer>
-                            </transformers>
-                            <createDependencyReducedPom>true</createDependencyReducedPom>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-    <dependencies>
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>opennms-dao-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.opennms.features.bsm</groupId>
-            <artifactId>org.opennms.features.bsm.persistence.api</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <!-- TODO MVR when merging with features/drift, comment this in -->
-        <!--<dependency>-->
-            <!--<groupId>org.opennms.features.flows.classification.persistence</groupId>-->
-            <!--<artifactId>org.opennms.features.flows.classification.persistence.api</artifactId>-->
-            <!--<version>${project.version}</version>-->
-        <!--</dependency>-->
-    </dependencies>
-</project>
